LayoutTests/inspector/styles/show-user-agent-styles-expected.txt

Tests that user agent styles are shown based on the setting: https://bugs.webkit.org/show_bug.cgi?id=37766.

I'm red.

I'm also red.


Running: testShowUserAgentStyles
WebInspector.settings.showUserAgentStyles = true
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
h1  { (show-user-agent-styles.html:5)
background-color: red;

[expanded] 
h1  { (user agent stylesheet)
display: block;
font-size: 2em;
-webkit-margin-before: 0.67em;
-webkit-margin-after: 0.67em;
-webkit-margin-start: 0px;
-webkit-margin-end: 0px;
font-weight: bold;



Running: testShowUserAgentStyles
WebInspector.settings.showUserAgentStyles = false
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
h1  { (show-user-agent-styles.html:5)
background-color: red;
